requirements.txt
django==3.0
flake8==3.7.0
pytest-django==3.9.0
six==1.14.0
black==22.3.0